refactor: split ShotEntryModal into form sections
